ISPN-7847 Move MergeOnStore to o.i.container (non-public package)
