+++ core/src/main/java/org/infinispan/statetransfer/CommitManager.java
@@ -7,7 +7,7 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
-import org.infinispan.atomic.MergeOnStore;
+import org.infinispan.container.MergeOnStore;
 import org.infinispan.container.DataContainer;
 import org.infinispan.container.entries.CacheEntry;
 import org.infinispan.context.Flag;
